requirements.txt
gradio>=4.0.0
transformers>=4.50.0
torch>=2.2.0
torchvision
accelerate
pydicom>=2.4.0
Pillow>=10.0.0
numpy>=1.24.0,<2.0
python-dotenv>=1.0.0